% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl.R
\name{parse_text}
\alias{parse_text}
\alias{parse_file}
\title{Alternative interfaces for [base::parse()]}
\usage{
parse_text(x)

parse_file(x)
}
\arguments{
\item{x}{For `parse_text`, the text to parse; for `parse_file` the file to parse.}
}
\description{
Alternative interfaces for [base::parse()]
}
