% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{py_parse_text}
\alias{py_parse_text}
\alias{py_parse_file}
\title{Alternative parser for R code with identations}
\usage{
py_parse_text(text)

py_parse_file(file)
}
\arguments{
\item{text}{character vector.  The text to parse.  Elements are treated
    as if they were lines of a file.  Other \R objects will be coerced
    to character if possible.}

\item{file}{a \link{connection}, or a character string giving the name of a
    file or a URL to read the expressions from.
    If \code{file} is \code{""} and \code{text} is missing or \code{NULL}
    then input is taken from the console.}
}
\description{
Alternative parser for R code with identations
}
