% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl.R
\name{repl}
\alias{repl}
\title{Launch a Read Eval Print Loop}
\usage{
repl(parser = parse_text, envir = parent.frame())
}
\arguments{
\item{parser}{the parsing function to use}

\item{envir}{environment to perform evaluations in}
}
\description{
Launch a Read Eval Print Loop
}
