% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s-exp-parser.R
\name{sexp_parse_file}
\alias{sexp_parse_file}
\alias{sexp_parse_text}
\title{Alternative S-expression parser for R code}
\usage{
sexp_parse_file(file)

sexp_parse_text(text)
}
\arguments{
\item{file}{a \link{connection}, or a character string giving the name of a
    file or a URL to read the expressions from.
    If \code{file} is \code{""} and \code{text} is missing or \code{NULL}
    then input is taken from the console.}

\item{text}{character vector.  The text to parse.  Elements are treated
    as if they were lines of a file.  Other \R objects will be coerced
    to character if possible.}
}
\description{
Alternative S-expression parser for R code
}
