% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl.R
\name{src}
\alias{src}
\title{Source a file with an alternative parser}
\usage{
src(dir, files = NULL, parser = parse_file,
  envir = asNamespace(packageName()))
}
\arguments{
\item{parser}{The parser to use, should take the file to parse as the first parameter}

\item{envir}{The environment to evaluate the parsed code in}

\item{file}{The file to parse}
}
\description{
Source a file with an alternative parser
}
