% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{tidy_parser}
\alias{tidy_parser}
\title{Alternative parser for R code}
\usage{
tidy_parser(text)
}
\arguments{
\item{text}{character vector.  The text to parse.  Elements are treated
    as if they were lines of a file.  Other \R objects will be coerced
    to character if possible.}
}
\description{
Alternative parser for R code
}
