% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl.R
\name{src}
\alias{src}
\title{Source a file with an alternative parser}
\usage{
src(file, parser = parse_file, envir = parent.frame())
}
\arguments{
\item{file}{The file to parse}

\item{parser}{The parser to use, should take the file to parse as the first parameter}

\item{envir}{The environment to evaluate the parsed code in}
}
\description{
Source a file with an alternative parser
}
