% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl.R
\name{src}
\alias{src}
\title{Source a file with an alternative parser}
\usage{
src(dir, package, parser = parse_file, files = NULL,
  envir = asNamespace(package))
}
\arguments{
\item{dir}{The installed directory from which to parse files}

\item{package}{The package}

\item{parser}{The parser to use, should take the file to parse as the first parameter}

\item{files}{The files to parse}

\item{envir}{The environment to evaluate the parsed code in}
}
\description{
Source a file with an alternative parser
}
